modularized the classroomsettings.tsx
