+++ lexilearner/app/classroom/[id]/classroomsettings.tsx
@@ -1,36 +1,22 @@
-import React, { useState, useRef, useEffect } from "react";
-import { Text } from "@/components/ui/text";
-import {
-  ScrollView,
-  View,
-  KeyboardAvoidingView,
-  Platform,
-  TouchableOpacity,
-  ActivityIndicator,
-  Keyboard,
-} from "react-native";
-import { Button } from "@/components/ui/button";
-import BackHeader from "@/components/BackHeader";
-import { Input } from "@/components/ui/input";
+import React, { useState } from "react";
+import { View } from "react-native";
+import { useUserStore } from "@/stores/userStore";
 import { useClassroomStore } from "@/stores/classroomStore";
-import { TextArea } from "@/components/ui/textarea";
 import { useMutation, useQueryClient, useQuery } from "@tanstack/react-query";
-import SearchBar, { SearchBarRef } from "@/components/Classroom/SearchBar";
 import {
   editClassroom as apiEditClassroom,
   deleteClassroom as apiDeleteClassroom,
   addPupilToClassroom as apiAddPupilToClassroom,
   searchPupils as apiSearchPupils,
   getPupilsFromClassroom,
-  Pupil,
   leaveClassroom as apiLeaveClassroom,
 } from "@/services/ClassroomService";
-import { router } from "expo-router";
-import { Ionicons, MaterialIcons } from "@expo/vector-icons";
-import { useUserStore } from "@/stores/userStore";
+import TeacherSetting from "@/components/Classroom/TeacherSetting";
+import PupilSetting from "@/components/Classroom/PupilSetting";
 
 export default function ClassroomSettings() {
   const queryClient = useQueryClient();
+
   const selectedClassroom = useClassroomStore(
     (state) => state.selectedClassroom
   );
@@ -38,22 +24,7 @@ export default function ClassroomSettings() {
     (state) => state.setSelectedClassroom
   );
 
-  const [editClassroomForm, setEditClassroomForm] = useState({
-    name: selectedClassroom?.name || "",
-    description: selectedClassroom?.description || "",
-  });
-
-  const [searchText, setSearchText] = useState("");
-  const [filteredPupils, setFilteredPupils] = useState<Pupil[]>([]);
-  const [showAddPupilModal, setShowAddPupilModal] = useState(false);
-  const [searching, setSearching] = useState(false);
-  const [showCurrentPupils, setShowCurrentPupils] = useState(false);
-  const [selectedPupils, setSelectedPupils] = useState<Pupil[]>([]);
-
-  const searchInputRef = useRef<SearchBarRef>(null);
   const user = useUserStore((state) => state.user);
-  const debounceTimerRef = useRef<NodeJS.Timeout | null>(null);
-  const isClearing = useRef(false);
 
   const { data: enrolledPupils, isLoading: loadingPupils } = useQuery({
     queryKey: ["classroomPupils", selectedClassroom?.id],
@@ -63,98 +34,6 @@ export default function ClassroomSettings() {
         : Promise.resolve([]),
     enabled: !!selectedClassroom?.id,
   });
-
-  useEffect(() => {
-    if (showAddPupilModal && searchInputRef.current) {
-      setTimeout(() => {
-        if (searchInputRef.current) {
-          searchInputRef.current.focus();
-        }
-      }, 100);
-    }
-  }, [showAddPupilModal]);
-
-  // Debounced search effect
-  useEffect(() => {
-    if (debounceTimerRef.current) {
-      clearTimeout(debounceTimerRef.current);
-    }
-
-    // Special handling for empty field - ensure keyboard stays visible
-    if (searchText.length === 0 && !isClearing.current) {
-      setTimeout(() => {
-        if (searchInputRef.current) {
-          searchInputRef.current.focus();
-        }
-      }, 100);
-    }
-
-    if (!showAddPupilModal) {
-      setFilteredPupils([]);
-      return;
-    }
-
-    // For empty or very short text, just clear results but keep keyboard
-    if (searchText.length <= 1) {
-      setFilteredPupils([]);
-      setSearching(false);
-      // Critical: Don't return early for empty text
-    }
-
-    if (searchText.length > 2) {
-      setSearching(true);
-    }
-
-    debounceTimerRef.current = setTimeout(async () => {
-      if (searchText.length <= 2) {
-        setSearching(false);
-        return;
-      }
-
-      try {
-        const results = await apiSearchPupils(searchText);
-
-        const filteredResults = results.filter(
-          (pupil) =>
-            !enrolledPupils?.some(
-              (enrolledPupil) => enrolledPupil.id === pupil.id
-            )
-        );
-
-        setFilteredPupils(filteredResults);
-
-        // Ensure keyboard stays up after query completes
-        setTimeout(() => {
-          if (searchInputRef.current && !isClearing.current) {
-            searchInputRef.current.focus();
-          }
-        }, 200);
-      } catch (error) {
-        console.error("Error searching pupils:", error);
-        const mockResults: Pupil[] = [
-          { id: "1", firstName: searchText, lastName: "Smith" },
-          { id: "2", firstName: searchText, lastName: "Johnson" },
-          { id: "3", firstName: searchText, lastName: "Williams" },
-        ];
-        setFilteredPupils(mockResults);
-
-        // Ensure keyboard stays up after query completes
-        setTimeout(() => {
-          if (searchInputRef.current && !isClearing.current) {
-            searchInputRef.current.focus();
-          }
-        }, 200);
-      } finally {
-        setSearching(false);
-      }
-    }, 1000);
-
-    return () => {
-      if (debounceTimerRef.current) {
-        clearTimeout(debounceTimerRef.current);
-      }
-    };
-  }, [searchText, enrolledPupils, showAddPupilModal]);
 
   const { mutateAsync: editClassroomMutation } = useMutation({
     mutationFn: ({
@@ -199,528 +78,23 @@ export default function ClassroomSettings() {
       });
     },
   });
-
-  const handleSelectPupil = (pupil: Pupil) => {
-    if (selectedPupils.some((p) => p.id === pupil.id)) {
-      setSelectedPupils(selectedPupils.filter((p) => p.id !== pupil.id));
-    } else {
-      setSelectedPupils([...selectedPupils, pupil]);
-    }
-  };
-
-  const handleRemoveSelectedPupil = (pupilId: string) => {
-    setSelectedPupils(selectedPupils.filter((p) => p.id !== pupilId));
-  };
-
-  const handleAddSelectedPupils = async () => {
-    if (selectedClassroom?.id && selectedPupils.length > 0) {
-      try {
-        // Intentionally dismiss keyboard here
-        isClearing.current = true;
-        Keyboard.dismiss();
-
-        for (const pupil of selectedPupils) {
-          await addPupilMutation({
-            classroomId: selectedClassroom.id,
-            pupilId: pupil.id || "",
-          });
-        }
-
-        setSelectedPupils([]);
-        setSearchText("");
-        setShowAddPupilModal(false);
-        console.log("Pupils added successfully");
-
-        setTimeout(() => {
-          isClearing.current = false;
-        }, 300);
-      } catch (error) {
-        console.error("Error adding pupils:", error);
-      }
-    }
-  };
-
-  const displayPupilName = (pupil: Pupil): string => {
-    return `${pupil.firstName} ${pupil.lastName}`;
-  };
-
-  const getRandomColor = (name: string) => {
-    const colors = [
-      "#845EC2",
-      "#D65DB1",
-      "#FF6F91",
-      "#FF9671",
-      "#FFC75F",
-      "#F9F871",
-      "#0089BA",
-      "#2C73D2",
-    ];
-    const hash = name.split("").reduce((acc, char) => {
-      return char.charCodeAt(0) + acc;
-    }, 0);
-    return colors[hash % colors.length];
-  };
-
-  const handleTextChange = (text: string) => {
-    setSearchText(text);
-  };
-
-  const AddPupilModal = () => {
-    const isClosingModal = useRef(false);
-
-    const handleCloseModal = () => {
-      isClosingModal.current = true;
-      isClearing.current = true;
-      Keyboard.dismiss();
-      setSearchText("");
-      setSelectedPupils([]);
-      setShowAddPupilModal(false);
-
-      setTimeout(() => {
-        isClearing.current = false;
-      }, 300);
-    };
-
-    return (
-      <KeyboardAvoidingView
-        behavior={Platform.OS === "ios" ? "padding" : "height"}
-        keyboardVerticalOffset={Platform.OS === "ios" ? 64 : 0}
-        style={{
-          position: "absolute",
-          top: 0,
-          left: 0,
-          right: 0,
-          bottom: 0,
-          backgroundColor: "white",
-          zIndex: 50,
-          height: "100%",
-        }}
-        enabled={true}
-      >
-        <View className="flex-1 bg-white">
-          {/* Header with Cancel, title, and Add button */}
-          <View className="relative flex-row items-center justify-center p-4 border-b border-gray-200">
-            {/* Cancel Button (Left) */}
-            <TouchableOpacity
-              onPress={handleCloseModal}
-              hitSlop={{ top: 10, bottom: 10, left: 10, right: 10 }}
-              className="absolute left-4"
-            >
-              <Text className="text-blue-500 text-lg">Cancel</Text>
-            </TouchableOpacity>
-
-            {/* Center Title */}
-            <Text className="text-black text-lg font-semibold">
-              Choose pupil
-            </Text>
-
-            {/* Add Button (Right) */}
-            {selectedPupils.length > 0 && (
-              <TouchableOpacity
-                onPress={handleAddSelectedPupils}
-                hitSlop={{ top: 10, bottom: 10, left: 10, right: 10 }}
-                className="absolute right-4"
-              >
-                <Text className="text-lg text-black">Add</Text>
-              </TouchableOpacity>
-            )}
-          </View>
-
-          {/* SearchBar component */}
-          <View className="p-2 px-4 mb-3">
-            <SearchBar
-              placeholder="Search"
-              initialValue={searchText}
-              onSearch={handleTextChange}
-              ref={searchInputRef}
-              debounceTime={1000}
-              blurOnSubmit={false}
-              isSearching={searching}
-              onClearPress={() => {
-                // Explicitly dismiss keyboard when X is pressed
-                isClearing.current = true;
-                Keyboard.dismiss();
-                setSearchText("");
-
-                // Reset flag after delay
-                setTimeout(() => {
-                  isClearing.current = false;
-                }, 300);
-              }}
-            />
-          </View>
-
-          {/* Selected Pupils */}
-          <View>
-            {selectedPupils.length > 0 && (
-              <View style={{ marginBottom: 5 }}>
-                <ScrollView
-                  horizontal
-                  className="pt-1 pb-1 px-2"
-                  showsHorizontalScrollIndicator={false}
-                  keyboardShouldPersistTaps="always"
-                  keyboardDismissMode="on-drag"
-                >
-                  {selectedPupils.map((pupil) => {
-                    const avatarColor = getRandomColor(
-                      pupil.firstName + pupil.lastName
-                    );
-                    const initials = `${pupil.firstName?.charAt(0) || ""}${
-                      pupil.lastName?.charAt(0) || ""
-                    }`;
-
-                    return (
-                      <View key={pupil.id} className="items-center mx-2 w-20">
-                        <View className="relative">
-                          <View
-                            className="h-16 w-16 rounded-full overflow-hidden justify-center items-center"
-                            style={{ backgroundColor: avatarColor }}
-                          >
-                            <Text className="text-white font-bold text-lg">
-                              {initials.toUpperCase()}
-                            </Text>
-                          </View>
-                          <TouchableOpacity
-                            style={{
-                              position: "absolute",
-                              top: -4,
-                              right: -4,
-                              backgroundColor: "grey",
-                              borderRadius: 999,
-                              padding: 4,
-                              zIndex: 10,
-                            }}
-                            onPress={() =>
-                              handleRemoveSelectedPupil(pupil.id || "")
-                            }
-                            hitSlop={{
-                              top: 10,
-                              bottom: 10,
-                              left: 10,
-                              right: 10,
-                            }}
-                          >
-                            <Ionicons name="close" size={14} color="black" />
-                          </TouchableOpacity>
-                        </View>
-                        <Text
-                          className="text-black text-center mt-1 text-xs"
-                          numberOfLines={1}
-                          ellipsizeMode="tail"
-                        >
-                          {pupil.firstName}
-                        </Text>
-                        <Text
-                          className="text-black text-center text-xs"
-                          numberOfLines={1}
-                          ellipsizeMode="tail"
-                        >
-                          {pupil.lastName}
-                        </Text>
-                      </View>
-                    );
-                  })}
-                </ScrollView>
-                <View />
-              </View>
-            )}
-
-            {/* Suggested Label */}
-            <Text
-              className="text-gray-500 text-lg font-medium px-4 mt-0"
-              style={{ marginBottom: 5 }}
-            >
-              Suggested
-            </Text>
-          </View>
-
-          {/* Pupil List - using keyboardDismissMode="on-drag" to allow dismissal on scroll */}
-          {searching ? (
-            <View className="flex-1 justify-center items-center">
-              <ActivityIndicator size="large" color="#0A84FF" />
-            </View>
-          ) : (
-            <ScrollView
-              className="flex-1"
-              keyboardShouldPersistTaps="handled"
-              keyboardDismissMode="on-drag"
-            >
-              {filteredPupils.length > 0 ? (
-                <>
-                  {filteredPupils.map((pupil) => {
-                    const avatarColor = getRandomColor(
-                      pupil.firstName + pupil.lastName
-                    );
-                    const initials = `${pupil.firstName?.charAt(0) || ""}${
-                      pupil.lastName?.charAt(0) || ""
-                    }`;
-                    const isSelected = selectedPupils.some(
-                      (p) => p.id === pupil.id
-                    );
-
-                    return (
-                      <View key={pupil.id}>
-                        <TouchableOpacity
-                          className="flex-row items-center justify-between py-3 px-3"
-                          onPress={() => handleSelectPupil(pupil)}
-                        >
-                          <View className="flex-row items-center">
-                            <View
-                              className="h-12 w-12 rounded-full overflow-hidden mr-3 justify-center items-center"
-                              style={{ backgroundColor: avatarColor }}
-                            >
-                              <Text className="text-white font-bold text-base">
-                                {initials.toUpperCase()}
-                              </Text>
-                            </View>
-                            <Text className="text-black text-base font-medium">
-                              {displayPupilName(pupil)}
-                            </Text>
-                          </View>
-                          <View
-                            className={`h-7 w-7 rounded-full border ${
-                              isSelected
-                                ? "bg-black-500 border-black-500 items-center justify-center"
-                                : "border-gray-400"
-                            }`}
-                          >
-                            {isSelected && (
-                              <Ionicons
-                                name="checkmark"
-                                size={18}
-                                color="black"
-                              />
-                            )}
-                          </View>
-                        </TouchableOpacity>
-                        <View
-                          className="flex-row"
-                          style={{ paddingLeft: 67, paddingRight: 8 }}
-                        >
-                          <View className="flex-1 border-b border-gray-200"></View>
-                        </View>
-                      </View>
-                    );
-                  })}
-                </>
-              ) : (
-                <>
-                  {searchText.length > 1 && !searching && (
-                    <View className="p-4 justify-center items-center">
-                      <Text className="text-gray-500">
-                        No matching pupils found
-                      </Text>
-                    </View>
-                  )}
-                </>
-              )}
-            </ScrollView>
-          )}
-        </View>
-      </KeyboardAvoidingView>
-    );
-  };
-
-  function TeacherSettings() {
-    return (
-      <View className="flex-1">
-        <ScrollView className="bg-background p-8">
-          <BackHeader />
-
-          <View className="py-1">
-            <Text className="text-[22px] text-center m-5 font-bold">
-              Classroom
-            </Text>
-            <Text className="font-bold">Classroom Name</Text>
-            <View>
-              <Input
-                className="border border-gray-300 rounded-md p-2 mt-2"
-                placeholder={editClassroomForm.name}
-                value={editClassroomForm.name}
-                onChangeText={(value: string) =>
-                  setEditClassroomForm({ ...editClassroomForm, name: value })
-                }
-              />
-            </View>
-          </View>
-          <TextArea
-            placeholder="Classroom Description..."
-            value={editClassroomForm.description}
-            onChangeText={(value: string) =>
-              setEditClassroomForm({
-                ...editClassroomForm,
-                description: value,
-              })
-            }
-          />
-
-          {/* Add Pupil Button */}
-          <TouchableOpacity
-            className="mt-6 mb-2 p-3 bg-blue-500 rounded-md flex-row justify-left items-center"
-            onPress={() => setShowAddPupilModal(true)}
-          >
-            <Ionicons name="person-add-outline" size={20} color="black" />
-            <Text className="text-black font-semibold ml-2">Add Pupils</Text>
-          </TouchableOpacity>
-
-          {/* Current Pupils */}
-          <View className="py-4 mt-4">
-            <TouchableOpacity
-              className="flex-row justify-between items-center border border-gray-300 rounded-md p-3"
-              onPress={() => setShowCurrentPupils(!showCurrentPupils)}
-            >
-              <Text className="font-bold text-black">Current Pupils</Text>
-              <MaterialIcons
-                name={
-                  showCurrentPupils
-                    ? "keyboard-arrow-up"
-                    : "keyboard-arrow-down"
-                }
-                size={24}
-                color="#333"
-              />
-            </TouchableOpacity>
-
-            {showCurrentPupils ? (
-              <View className="border-x border-b border-gray-300 rounded-b-md bg-white">
-                {loadingPupils ? (
-                  <View className="p-4 items-center">
-                    <ActivityIndicator size="small" color="#0000ff" />
-                  </View>
-                ) : enrolledPupils && enrolledPupils.length > 0 ? (
-                  <>
-                    {enrolledPupils.map((pupil) => {
-                      const avatarColor = getRandomColor(
-                        pupil.firstName + pupil.lastName
-                      );
-                      const initials = `${pupil.firstName?.charAt(0) || ""}${
-                        pupil.lastName?.charAt(0) || ""
-                      }`;
-
-                      return (
-                        <View
-                          key={pupil.id}
-                          className="p-3 border-b border-gray-200 flex-row items-center"
-                        >
-                          <View
-                            className="h-8 w-8 rounded-full overflow-hidden mr-3 justify-center items-center"
-                            style={{ backgroundColor: avatarColor }}
-                          >
-                            <Text className="text-white font-bold text-xs">
-                              {initials.toUpperCase()}
-                            </Text>
-                          </View>
-                          <Text className="text-black">
-                            {displayPupilName(pupil)}
-                          </Text>
-                        </View>
-                      );
-                    })}
-                  </>
-                ) : (
-                  <View className="p-3">
-                    <Text className="text-gray-500 italic">
-                      No pupils enrolled in this classroom yet
-                    </Text>
-                  </View>
-                )}
-              </View>
-            ) : null}
-          </View>
-
-          <View className="p-5 bottom-0">
-            <Button className="m-5 bg-yellow-500" onPress={() => {}}>
-              <Text className="text-black font-semibold">
-                Generate Classroom Report
-              </Text>
-            </Button>
-            <Button
-              className="mx-5 bg-yellow-500"
-              onPress={async () => {
-                if (selectedClassroom?.id) {
-                  try {
-                    await editClassroomMutation({
-                      classroomForm: editClassroomForm,
-                      classroomId: selectedClassroom.id,
-                    });
-
-                    setSelectedClassroom({
-                      ...selectedClassroom,
-                      ...editClassroomForm,
-                    });
-
-                    router.back();
-                    console.log("Classroom edited successfully");
-                  } catch (error) {
-                    console.error("Error editing classroom:", error);
-                  }
-                } else {
-                  console.error("Classroom ID is not available");
-                }
-              }}
-            >
-              <Text className="text-black font-semibold">Edit Classroom</Text>
-            </Button>
-            <Button
-              className="bg-red-500 m-5"
-              onPress={async () => {
-                if (selectedClassroom?.id) {
-                  try {
-                    await deleteClassroomMutation({
-                      classroomId: selectedClassroom.id,
-                    });
-                    setSelectedClassroom(null);
-                    router.push("/classroom");
-                  } catch (error) {
-                    console.error("Error deleting classroom:", error);
-                  }
-                }
-              }}
-            >
-              <Text className="text-white font-semibold">Delete Classroom</Text>
-            </Button>
-          </View>
-        </ScrollView>
-
-        {showAddPupilModal && <AddPupilModal />}
-      </View>
-    );
-  }
-
-  function PupilSettings() {
-    return (
-      <View className="flex-1">
-        <ScrollView className="bg-background p-8">
-          <BackHeader />
-          <View className="py-1">
-            <Text className="text-[22px] text-center m-5 font-bold">
-              Classroom
-            </Text>
-          </View>
-          <View className="p-5 bottom-0 my-5">
-            <Button
-              className="m-5 bg-orange"
-              onPress={async () => {
-                if (selectedClassroom?.id) {
-                  try {
-                    await leaveClassroomMutation({
-                      classroomId: selectedClassroom?.id,
-                    });
-                    setSelectedClassroom(null);
-                    router.replace("/classroom");
-                  } catch (error) {
-                    console.error("Error leaving classroom:", error);
-                  }
-                }
-              }}
-            >
-              <Text className="text-black font-semibold">Leave Classroom</Text>
-            </Button>
-          </View>
-        </ScrollView>
-      </View>
-    );
-  }
-
-  return user?.role === "Teacher" ? <TeacherSettings /> : <PupilSettings />;
+  // buwag nani ha
+  return user?.role === "Teacher" ? (
+    <TeacherSetting
+      selectedClassroom={selectedClassroom}
+      setSelectedClassroom={setSelectedClassroom}
+      enrolledPupils={enrolledPupils}
+      loadingPupils={loadingPupils}
+      editClassroomMutation={editClassroomMutation}
+      deleteClassroomMutation={deleteClassroomMutation}
+      addPupilMutation={addPupilMutation}
+      apiSearchPupils={apiSearchPupils}
+    />
+  ) : (
+    <PupilSetting
+      selectedClassroom={selectedClassroom}
+      setSelectedClassroom={setSelectedClassroom}
+      leaveClassroomMutation={leaveClassroomMutation}
+    />
+  );
 }
